experiments
moving variables to int64 and replacing atoi/atoll with Util::safe_atoi64

+++ daemon/extension/QueueScript.cpp
@@ -58,7 +58,7 @@ private:
 	int m_priority;
 	CString m_dupeKey;
 	EDupeMode m_dupeMode;
-	int m_dupeScore;
+	int64 m_dupeScore;
 	NzbParameterList m_parameters;
 	int m_prefixLen;
 	std::shared_ptr<const Extension::Script> m_script;
@@ -391,7 +391,7 @@ bool QueueScriptCoordinator::UsableScript(std::shared_ptr<const Extension::Scrip
 	return false;
 }
 
-NzbInfo* QueueScriptCoordinator::FindNzbInfo(DownloadQueue* downloadQueue, int nzbId)
+NzbInfo* QueueScriptCoordinator::FindNzbInfo(DownloadQueue* downloadQueue, int64 nzbId)
 {
 	NzbInfo* nzbInfo = downloadQueue->GetQueue()->Find(nzbId);
 	if (nzbInfo)
@@ -463,7 +463,7 @@ void QueueScriptCoordinator::CheckQueue()
 	}
 }
 
-bool QueueScriptCoordinator::HasJob(int nzbId, bool* active)
+bool QueueScriptCoordinator::HasJob(int64 nzbId, bool* active)
 {
 	Guard guard(m_queueMutex);
